refactor: merge test
Merge SkipFailingEventsTest with EngineErrorHandlingTest, to make more clear what they actually testing
and in which responsibility this test lies.
